doit: Split out cmd_reset
